Switch to new, more updated user agents list
The user agent list at user-agents.org has not been updated in a while.
So we're making the switch to
https://github.com/monperrus/crawler-user-agents.

Unfortunately, this new list does not categorize user agents. So the
following methods have been removed: #browser?, #checker?, #downloader?,
 #proxy?, #crawler? and #spam?.

This commit should cause a major version bump since it contains breaking
changes to the API.